fixing filtering through notebook; reorganizing integration tests
