+++ .github/workflows/tests.yml
@@ -35,4 +35,4 @@ jobs:
 
       - name: Pytest
         run: |
-          pytest -v -n0
+          pytest -v -n0 --dist=loadscope
